Move hook and ProgressBar for readability
