+++ src/progress_bar_solution/ProgressBar/ProgressBar.js
