+++ src/progress_bar_solution/ProgressBar/ProgressBar.test.js
