+++ src/progress_bar_solution/ProgressBar/index.js
