+++ src/progress_bar_solution/ProgressBar/progress-bar.css
