+++ src/progress_bar_solution/hooks/constants.js
